Use the System Face Detection
Use the AVCaptureMetaDataOutput for face detection on the AVCaptureSession. This is amazingly fast in comparison to the Dlib face detector. The resulting bounds rectangle is then fed into Dlib for landmarking, which does not seem to need much time at all. 

The MetaDataOutput returns a scaled rectangle with values between 0 and 1, which we need to convert and blow up to the size of the image.

Current limitation: It only applies landmarking to the first face found in an image.